Address comments.
1. Move ModuleEntry to module.py
2. Reuse org.apache.flink.table.utils.ModuleMock
