+++ flink-python/pyflink/table/module.py
@@ -17,7 +17,7 @@
 ################################################################################
 from pyflink.java_gateway import get_gateway
 
-__all__ = ['Module', 'HiveModule']
+__all__ = ['HiveModule', 'Module', 'ModuleEntry']
 
 
 class Module(object):
@@ -49,3 +49,37 @@ class HiveModule(Module):
         else:
             j_hive_module = gateway.jvm.org.apache.flink.table.module.hive.HiveModule(hive_version)
         super(HiveModule, self).__init__(j_hive_module)
+
+
+class ModuleEntry(object):
+    """
+    A POJO to represent a module's name and use status.
+    """
+
+    def __init__(self, name: str, used: bool, j_module_entry=None):
+        if j_module_entry is None:
+            gateway = get_gateway()
+            self._j_module_entry = gateway.jvm.org.apache.flink.table.module.ModuleEntry(name, used)
+        else:
+            self._j_module_entry = j_module_entry
+
+    def name(self) -> str:
+        return self._j_module_entry.name()
+
+    def used(self) -> bool:
+        return self._j_module_entry.used()
+
+    def __repr__(self):
+        return self._j_module_entry.toString()
+
+    def __eq__(self, other):
+        return isinstance(other, self.__class__) and self._j_module_entry == other._j_module_entry
+
+    def __hash__(self):
+        return self._j_module_entry.hashCode()
+
+    def __ne__(self, other):
+        return not self.__eq__(other)
+
+    def __lt__(self, other):
+        return not self.__eq__(other)
